Add dynamodb stream trigger for notifier
In place of SNS topics, for now I will do the event filtering at the
notifier. This will only trigger the lambda for the few areas that I
care about. In the future, I may instead put SNS in-between DynamoDB and
the notifier lambda but that's more work than I want to do for now.